Move extension version check to common workspace
