+++ packages/app/app/scripts/desktop/app/globals.ts
@@ -3,7 +3,7 @@ import '../browser/browser-init';
 import { webcrypto } from 'node:crypto';
 import { Headers } from 'node-fetch';
 import setupSentry from '../../lib/setupSentry';
-import { getVersion } from '../utils/version';
+import { getDesktopVersion } from '../utils/version';
 
 declare const global: typeof globalThis & {
   stateHooks: Record<string, any>;
@@ -42,7 +42,7 @@ global.stateHooks = {};
 
 // setup sentry error reporting
 global.sentry = setupSentry({
-  release: getVersion(),
+  release: getDesktopVersion(),
   getState: () => global.stateHooks?.getSentryState?.() || {},
 });
