+++ packages/app/lavamoat/browserify/flask/policy.json
@@ -5883,7 +5883,26 @@
     "external:../common/dist/index.js": {
       "packages": {
         "external:../common/dist/browser/index.js": true,
+        "external:../common/dist/version-check.js": true,
         "external:../common/dist/web-socket-stream.js": true
+      }
+    },
+    "external:../common/dist/utils/config.js": {
+      "packages": {
+        "browserify>process": true
+      }
+    },
+    "external:../common/dist/utils/stream.js": {
+      "packages": {
+        "browserify>stream-browserify": true,
+        "external:../common/dist/constants.js": true
+      }
+    },
+    "external:../common/dist/version-check.js": {
+      "packages": {
+        "external:../common/dist/utils/config.js": true,
+        "external:../common/dist/utils/stream.js": true,
+        "external:../common/node_modules/loglevel/lib/loglevel.js": true
       }
     },
     "external:../common/dist/web-socket-stream.js": {
