+++ packages/common/src/index.ts
@@ -1,2 +1,3 @@
 export { browser } from './browser';
 export * from './web-socket-stream';
+export * from './version-check';
